+++ src/main/java/com/fasterxml/jackson/databind/util/concurrentlinkedhashmap/Weigher.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fasterxml.jackson.databind.concurrentlinkedhashmap;
+package com.fasterxml.jackson.databind.util.concurrentlinkedhashmap;
 
 import javax.annotation.concurrent.ThreadSafe;
